agentverse/requirements.txt
fetchai
python-dotenv
uagents
uagents-core
requests
